refactor wb styles
